Update tests to ensure they're testing what we actually want to test.
The test_image_changed_detection() has been updated to check two conditions (which will be added to the actual system in 0.2.0).

The test_textureviews_changed_detection() has been enabled, as I've figured out how to 'test' it. While it doesn't actually change anything (since doing so is...difficult) we can somewhat rely on Bevy's internal behavior to ensure that this *should*, theoretically, work.